+++ src/main/java/com/github/antag99/retinazer/EngineService.java
@@ -1,0 +1,7 @@
+package com.github.antag99.retinazer;
+
+/** General {@link Engine} service.
+ * @see WireResolver
+ * @see EntitySystem */
+public interface EngineService {
+}