+++ src/test/java/com/github/antag99/retinazer/FlagComponentB.java
@@ -1,4 +1,0 @@
-package com.github.antag99.retinazer;
-
-public final class FlagComponentB implements Component {
-}