+++ src/test/java/com/github/antag99/retinazer/FlagComponentC.java
@@ -1,4 +1,0 @@
-package com.github.antag99.retinazer;
-
-public final class FlagComponentC implements Component {
-}